refactor helperFunction and fixed updating bug
