+++ client/components/categoryView.js
@@ -1,8 +1,9 @@
 import React from 'react'
 import {connect} from 'react-redux'
 import axios from 'axios'
-import {getCategoryPosts, getAllCategories} from '.././store/post.js'
+import {getCategoryPosts} from '.././store/post.js'
 import store from '.././store'
+import {formatDate} from '../helperFunctions'
 
 class CategoryView extends React.Component{
 
@@ -13,21 +14,11 @@ class CategoryView extends React.Component{
       })
       .catch(err => console.log(err))
     }
-
-    formatDate(time){
-        let yearMonthDate = time.slice(0, time.indexOf('T'))
-        let year = yearMonthDate.slice(0,4)
-        let month = yearMonthDate.slice(5, 7)
-        let date = yearMonthDate.slice(8)
-        let months = { '01': 'Jan', '02': 'Feb', '03': 'Mar', '04' : 'Apr', '05': 'May', '06' :'June', 
-        '07': 'July','08': 'Aug', '09': 'Sep', '10': 'Oct', '11': 'Nov', '12': 'Dec'}
-        return `Published on ${months[month]} ${date}, ${year}`
-     }
-
+      
    render(){
-       let formatDate = this.formatDate
        let postsDisplay = this.props.posts
        let categoryDisplay = this.props.categories
+       console.log(categoryDisplay, 'categoryDisplay in catview')
     return (
         <div>
         {!postsDisplay.length ? <p>There are no posts here yet, but there will be soon!</p> :
