+++ client/components/home.js
@@ -2,24 +2,30 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {getAllPosts} from '.././store/post.js'
+import store from '.././store'
 import axios from 'axios'
 import {Sidebar} from '../components'
+import {formatDate} from '../helperFunctions'
 /**
  * COMPONENT
  */
-export const Home = (props) => {
-  const {posts, categories} = props
-  let categoryDisplay = categories
+
+class  Home extends React.Component{
   
-  function formatDate(time){
-    let yearMonthDate = time.slice(0, time.indexOf('T'))
-    let year = yearMonthDate.slice(0,4)
-    let month = yearMonthDate.slice(5, 7)
-    let date = yearMonthDate.slice(8)
-    let months = { '01': 'Jan', '02': 'Feb', '03': 'Mar', '04' : 'Apr', '05': 'May', '06' :'June', 
-    '07': 'July','08': 'Aug', '09': 'Sep', '10': 'Oct', '11': 'Nov', '12': 'Dec'}
-    return `Published on ${months[month]} ${date}, ${year}`
- }
+  componentDidMount = () => {
+    axios.get('/get')
+    .then(res => {
+        store.dispatch(getAllPosts(res.data.info, res.data.categories))
+    })
+    .catch(err => console.log(err))
+   }
+  
+   render(){
+    const {posts, categories} = this.props
+    let categoryDisplay = categories
+  console.log(categoryDisplay, 'categoryDisplay')
+  console.log(posts, 'posts')
+
   return (
     <div>
       <Sidebar />
@@ -30,7 +36,7 @@ export const Home = (props) => {
           <p>{formatDate(post.createdAt)}</p>
           <div id="navcontainer">
                             <ul id="navlist">
-                            {categoryDisplay[index].tags.map(function(category){
+                            {categoryDisplay.length !== posts.length ? <p>" "</p> : categoryDisplay[index].tags.map(function(category){
                                 return <li key={post.id}>{category}</li>
                             })}
                             </ul>
@@ -39,6 +45,7 @@ export const Home = (props) => {
       })}
     </div>
   )
+  }
 }
 
 /**
@@ -51,14 +58,5 @@ const mapState = (state) => {
   }
 }
 
-const mapDispatch = (dispatch) => {
-  return ( dispatch =>
-   axios.get('/get')
-    .then(res => {
-        dispatch(getAllPosts(res.data.info, res.data.categories))
-    })
-    .catch(err => console.log(err)))
-   }
-
-export default connect(mapState, mapDispatch)(Home)
+export default connect(mapState)(Home)
